コンストラクタ引数 speed_sampling_time の単位を変更

+++ wheel_control.h
@@ -73,9 +73,9 @@ public:
     *
     * @param
     *   wheel_diameter     : milli meter [mm].
-    *   speed_sampling_time: milli second [ms].
+    *   speed_sampling_time: second [s].
     */
-    WheelControl(Motor &motor, Encoder &encoder, float32_t wheel_diameter, uint16_t speed_sampling_time);
+    WheelControl(Motor &motor, Encoder &encoder, float32_t wheel_diameter, float32_t speed_sampling_time);
 
     /**
      * @fn 呼び出し側のプログラムで，この関数をタイマ割り込み(任意の周期)で計測しないと動作しないので注意してください．
@@ -109,9 +109,9 @@ private:
     float32_t _speed;  //< [mm/s] mm_per_second.
     Encoder &_encoder;
     Motor &_motor;
-    const int32_t _speed_sampling_time;  // milli second [ms]
-    const float32_t _distance_per_pulse; // [mm/pulse] 1パルスにつき進む距離[mm]
-    const float32_t _speed_per_pulse;    // callbackされるサンプリングタイムも考慮したパラメータ
+    const float32_t _speed_sampling_time; // second [s]
+    const float32_t _distance_per_pulse;  // [mm/pulse] 1パルスにつき進む距離[mm]
+    const float32_t _speed_per_pulse;     // callbackされるサンプリングタイムも考慮したパラメータ
 };
 
 }  // namespace mslh
